COMPONENTS SEPARATION NEW FOLDERS ASIGNED FOR EACH COMPONENT USED IN THE APP
