+++ src/components/MainPage/index.js
@@ -1,0 +1,2 @@
+export {InputSearch} from './InputSearch.js';
+export {RecipeContainer} from './RecipeContainer.js';
