+++ src/components/RecipePage/index.js
@@ -1,0 +1,2 @@
+export {Description} from './Description.js';
+export {Cover} from './Cover.js';
