+++ src/screens/index.js
@@ -1,0 +1,2 @@
+export {MainPage} from './MainPage.js';
+export {RecipePage} from './RecipePage.js';
